restructure files, alias WalletsEvents as WalletsEventsListeners, add docs
