+++ packages/core/app/src/wallets.ts
@@ -7,50 +7,28 @@ import type {
     WindowAppReadyEventAPI,
 } from '@wallet-standard/base';
 
-/** TODO: docs */
-export interface Wallets {
-    /**
-     * TODO: docs
-     */
-    register(...wallets: Wallet[]): () => void;
-
-    /**
-     * TODO: docs
-     */
-    get(): ReadonlyArray<Wallet>;
-
-    /**
-     * TODO: docs
-     */
-    on<E extends WalletsEventNames = WalletsEventNames>(event: E, listener: WalletsEvents[E]): () => void;
-}
-
-/** TODO: docs */
-export interface WalletsEvents {
-    /**
-     * Emitted when wallets are registered.
-     *
-     * @param wallets Wallets that were registered.
-     */
-    register(...wallets: Wallet[]): void;
-
-    /**
-     * Emitted when wallets are unregistered.
-     *
-     * @param wallets Wallets that were unregistered.
-     */
-    unregister(...wallets: Wallet[]): void;
-}
-
-/** TODO: docs */
-export type WalletsEventNames = keyof WalletsEvents;
-
 let wallets: Wallets | undefined = undefined;
 const registered = new Set<Wallet>();
-const listeners: { [E in WalletsEventNames]?: WalletsEvents[E][] } = {};
+const listeners: { [E in WalletsEventNames]?: WalletsEventsListeners[E][] } = {};
 
 /**
- * TODO: docs
+ * Get an API for {@link Wallets.get | getting}, {@link Wallets.on | listening for}, and
+ * {@link Wallets.register | registering} {@link "@wallet-standard/base".Wallet | Wallets}.
+ *
+ * When called for the first time --
+ *
+ * This dispatches a {@link "@wallet-standard/base".WindowAppReadyEvent} to notify each Wallet that the app is ready
+ * to register it.
+ *
+ * This also adds a listener for {@link "@wallet-standard/base".WindowRegisterWalletEvent} to listen for a notification
+ * from each Wallet that the Wallet is ready to be registered by the app.
+ *
+ * This combination of event dispatch and listener guarantees that each Wallet will be registered synchronously as soon
+ * as the app is ready whether the app loads before or after each Wallet.
+ *
+ * @return API for getting, listening for, and registering Wallets.
+ *
+ * @group App
  */
 export function getWallets(): Wallets {
     if (wallets) return wallets;
@@ -74,6 +52,81 @@ export function getWallets(): Wallets {
     return wallets;
 }
 
+/**
+ * API for {@link Wallets.get | getting}, {@link Wallets.on | listening for}, and
+ * {@link Wallets.register | registering} {@link "@wallet-standard/base".Wallet | Wallets}.
+ *
+ * @group App
+ */
+export interface Wallets {
+    /**
+     * Get all Wallets that have been registered.
+     *
+     * @return Registered Wallets.
+     */
+    get(): readonly Wallet[];
+
+    /**
+     * Add an event listener and subscribe to events for Wallets that are
+     * {@link WalletsEventsListeners.register | registered} and
+     * {@link WalletsEventsListeners.unregister | unregistered}.
+     *
+     * @param event    Event type to listen for. {@link WalletsEventsListeners.register | `register`} and
+     * {@link WalletsEventsListeners.unregister | `unregister`} are the only event types.
+     * @param listener Function that will be called when an event of the type is emitted.
+     *
+     * @return
+     * `off` function which may be called to remove the event listener and unsubscribe from events.
+     *
+     * As with all event listeners, be careful to avoid memory leaks.
+     */
+    on<E extends WalletsEventNames>(event: E, listener: WalletsEventsListeners[E]): () => void;
+
+    /**
+     * Register Wallets. This can be used to programmatically wrap non-standard wallets as Standard Wallets.
+     *
+     * Apps generally do not need to, and should not, call this.
+     *
+     * @param wallets Wallets to register.
+     */
+    register(...wallets: Wallet[]): () => void;
+}
+
+/**
+ * Types of event listeners of the {@link Wallets} API.
+ *
+ * @group App
+ */
+export interface WalletsEventsListeners {
+    /**
+     * Emitted when Wallets are registered.
+     *
+     * @param wallets Wallets that were registered.
+     */
+    register(...wallets: Wallet[]): void;
+
+    /**
+     * Emitted when Wallets are unregistered.
+     *
+     * @param wallets Wallets that were unregistered.
+     */
+    unregister(...wallets: Wallet[]): void;
+}
+
+/**
+ * Names of {@link WalletsEventsListeners} that can be listened for.
+ *
+ * @group App
+ */
+export type WalletsEventNames = keyof WalletsEventsListeners;
+
+/**
+ * @deprecated Use {@link WalletsEventsListeners} instead.
+ *
+ * @group App
+ */
+export type WalletsEvents = WalletsEventsListeners;
+
 function register(...wallets: Wallet[]): () => void {
     // Filter out wallets that have already been registered.
     // This prevents the same wallet from being registered twice, but it also prevents wallets from being
@@ -92,11 +145,11 @@ function register(...wallets: Wallet[]): () => void {
     };
 }
 
-function get(): ReadonlyArray<Wallet> {
+function get(): readonly Wallet[] {
     return [...registered];
 }
 
-function on<E extends WalletsEventNames>(event: E, listener: WalletsEvents[E]): () => void {
+function on<E extends WalletsEventNames>(event: E, listener: WalletsEventsListeners[E]): () => void {
     listeners[event]?.push(listener) || (listeners[event] = [listener]);
     // Return a function that removes the event listener.
     return function off(): void {
@@ -148,7 +201,11 @@ class AppReadyEvent extends Event implements WindowAppReadyEvent {
     }
 }
 
-/** @deprecated */
+/**
+ * @deprecated Use {@link getWallets} instead.
+ *
+ * @group Deprecated
+ */
 export function DEPRECATED_getWallets(): Wallets {
     if (wallets) return wallets;
     wallets = getWallets();
