+++ packages/core/base/src/index.ts
@@ -1,2 +1,3 @@
+export * from './identifier.js';
 export * from './wallet.js';
 export * from './window.js';
