+++ packages/core/wallet/src/index.ts
@@ -1,3 +1,2 @@
-export * from './account.js';
 export * from './register.js';
 export * from './util.js';
