+++ packages/core/wallet/src/register.ts
@@ -6,6 +6,22 @@ import type {
     WindowRegisterWalletEventCallback,
 } from '@wallet-standard/base';
 
+/**
+ * Register a {@link "@wallet-standard/base".Wallet} as a Standard Wallet with the app.
+ *
+ * This dispatches a {@link "@wallet-standard/base".WindowRegisterWalletEvent} to notify the app that the Wallet is
+ * ready to be registered.
+ *
+ * This also adds a listener for {@link "@wallet-standard/base".WindowAppReadyEvent} to listen for a notification from
+ * the app that the app is ready to register the Wallet.
+ *
+ * This combination of event dispatch and listener guarantees that the Wallet will be registered synchronously as soon
+ * as the app is ready whether the Wallet loads before or after the app.
+ *
+ * @param wallet Wallet to register.
+ *
+ * @group Wallet
+ */
 export function registerWallet(wallet: Wallet): void {
     const callback: WindowRegisterWalletEventCallback = ({ register }) => register(wallet);
     try {
@@ -58,7 +74,11 @@ class RegisterWalletEvent extends Event implements WindowRegisterWalletEvent {
     }
 }
 
-/** @deprecated */
+/**
+ * @deprecated Use {@link registerWallet} instead.
+ *
+ * @group Deprecated
+ */
 export function DEPRECATED_registerWallet(wallet: Wallet): void {
     registerWallet(wallet);
     try {
